core_requirements.txt
django-rest-framework==0.1.0
mongoengine==0.19.1
pip==20.0.2
psycopg2==2.8.4
setuptools==46.1.3
wheel==0.34.2
